Linux git-lfs instructions used brew
The Linux instructions for installing "git lfs" used brew, which may be fine, but apt is used throughout the install instructions and brew is not.  Also, apt is native to the OS and brew is not.  So, I pulled the instructions from github and replaced brew (https://github.com/git-lfs/git-lfs/wiki/Installation).  

It is my opinion that dependencies should be very easy to install so developers spend their time writing code with IOST, not wrestling with the development environment.

Tested on a fresh install of Ubuntu 16.04LTS and 18.04LTS.